remove hard Devel::Peek dependency and lazy load for daemons
It's only useful for a corner case in long-running daemons when
an admin decides to compact or vacuum a Xapian or SQLite DB.

As a result, other scripts should run slightly faster.  For
instance, this saves about 80ms (2.710s => 2.630s) in t/mda.t
on my remote workstation.

While we're at it, make sure EvCleanup is properly require'd
in Daemon.pm and HTTP.pm and document our use of Devel::Peek.