wwwatomstream: fix per-feed <id>
We were emitting the same "<id>mailto:name@domain</id>" tag
for every feed (but not per-feed entry).  This could cause
feed readers to mistake the top (news.atom) feed for other
feeds (search results, or per-thread feeds).

This is technically a breaking change for people relying on
per-thread or per-query feeds, but the only alternative is
to remain broken for anybody trying to follow multiple feeds
off the same inbox.