Makefile.PL: allow overriding "prove" from make CLI
Development versions of Perl install "prove$VERSION"
where $VERSION is something like "5.31.7".  This makes
it easier to test everything we have against development
versions of Perl5.

Note: I could not find a way to get quoting right to use the
"--exec $(PERL)" option of prove(1), but that would be the best
option for working transparently after running:

	perl5.31.7 Makefile.PL