t/edit.t: drop redundant "delete local $ENV{...}"
"delete local" is only in Perl v5.11.0, and we only depend on
Perl v5.10.1.  We already localize and delete it as two separate
statements immediately above.

I wish this was hidden behind a "use feature" flag like other
new-fangled things :<.  Oh well, I think the oldest Perl actually
in use for this project is 5.16 (CentOS 7.x).