doc: recommend -compact after --reindex
It's likely a user will be low on space after running --reindex,
so recommend the use of public-inbox-compact afterwards.

And add a few more notes about using public-inbox-compact to
clarify it's for inboxes-only (and not any old Xapian DBs) that
using xapian-compact(1) directly is error-prone and likely to
break things.